[TASK] Finally implement TTL feature on cached assets, default life 3 hrs

+++ Classes/Service/AssetService.php
@@ -268,25 +268,27 @@ class Tx_Vhs_Service_AssetService implements t3lib_Singleton {
 	 * @return string
 	 */
 	private function writeCachedMergedFileAndReturnTag($assets, $type) {
-		// @todo Finish implementation. This line is currently not used. I guess the intention was to better handle cache time of a file (vs. writing the file every time)
-		$ttl = (TRUE === isset($GLOBALS['TSFE']->tmpl->setup['config.']['cache_period']) && $GLOBALS['TSFE']->tmpl->setup['config.']['cache_period'] !== 0);
+		$ttl = (TRUE === isset($GLOBALS['TSFE']->tmpl->setup['config.']['cache_period']) && $GLOBALS['TSFE']->tmpl->setup['config.']['cache_period'] !== 0)
+			? $GLOBALS['TSFE']->tmpl->setup['config.']['cache_period'] : 10800;
 		$source = '';
-		foreach ($assets as $name => $asset) {
-			$settings = $this->extractAssetSettings($asset);
-			if (TRUE === (isset($settings['namedChunks']) && 0 < $settings['namedChunks']) || FALSE === isset($settings['namedChunks'])) {
-				$source .= '/* ' . $name . ' */' . LF;
-			}
-			$source .= $this->extractAssetContent($asset) . LF;
-			// Put a return carriage between assets preventing broken content.
-			$source .= "\n";
-		}
 		$assetName = implode('-', array_keys($assets));
 		if ($GLOBALS['TSFE']->tmpl->setup['plugin.']['tx_vhs.']['assets.']['mergedAssetsUseHashedFilename']) {
 			$assetName = md5($assetName);
 		}
 		$fileRelativePathAndFilename = 'typo3temp/vhs-assets-' . $assetName . '.'.  $type;
 		$fileAbsolutePathAndFilename = t3lib_div::getFileAbsFileName($fileRelativePathAndFilename);
-		file_put_contents($fileAbsolutePathAndFilename, $source);
+		if (FALSE === file_exists($fileAbsolutePathAndFilename) || filectime($fileAbsolutePathAndFilename) < time() - $ttl) {
+			foreach ($assets as $name => $asset) {
+				$settings = $this->extractAssetSettings($asset);
+				if (TRUE === (isset($settings['namedChunks']) && 0 < $settings['namedChunks']) || FALSE === isset($settings['namedChunks'])) {
+					$source .= '/* ' . $name . ' */' . LF;
+				}
+				$source .= $this->extractAssetContent($asset) . LF;
+				// Put a return carriage between assets preventing broken content.
+				$source .= "\n";
+			}
+			file_put_contents($fileAbsolutePathAndFilename, $source);
+		}
 		$fileRelativePathAndFilename .= $this->appendModificationTime($fileRelativePathAndFilename);
 		$fileRelativePathAndFilename = $this->prefixPath($fileRelativePathAndFilename);
 		return $this->generateTagForAssetType($type, NULL, $fileRelativePathAndFilename);
